Fixed kafka and gateway configurations. Also finished junit testing and api testing in Postman

+++ course-ms/src/main/java/com/lms/coursems/repository/CourseRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CourseRepository extends JpaRepository<Course, Integer> {
+public interface CourseRepository extends JpaRepository<Course, Long> {
 
     Course findByNameAndDel(String courseName, Integer del);
